안쓰는 코드 삭제 및 리팩터링
