--- src/page/ChatScreen.jsx
@@ -1,7 +1,0 @@
-function ChatScreen() {
-    return (
-        <div>챗봇화면</div>
-    )
-}
-
-export default ChatScreen;
